ZOOKEEPER-2313 Refactor ZooKeeperServerBean and its subclasses
