+++ zookeeper-server/src/main/java/org/apache/zookeeper/server/quorum/ObserverBean.java
@@ -16,9 +16,12 @@
  * limitations under the License.
  */
 
-package org.apache.zookeeper.server;
+package org.apache.zookeeper.server.quorum;
 
 import java.net.InetSocketAddress;
+
+import org.apache.zookeeper.server.ZooKeeperServer;
+import org.apache.zookeeper.server.ZooKeeperServerBean;
 import org.apache.zookeeper.server.quorum.Observer;
 import org.apache.zookeeper.server.quorum.ObserverMXBean;
 import org.apache.zookeeper.server.quorum.QuorumPeer;
@@ -32,12 +35,8 @@ public class ObserverBean extends ZooKeeperServerBean implements ObserverMXBean 
     private Observer observer;
 
     public ObserverBean(Observer observer, ZooKeeperServer zks) {
-        super(zks);
+        super(zks, "ObserverServer_port" + zks.getClientPort());
         this.observer = observer;
-    }
-
-    public String getName() {
-        return "Observer";
     }
 
     public int getPendingRevalidationCount() {
